Crate script_plugins

Servo’s compiler plugin/macro crate

This crate provides the #[unrooted_must_root_lint::must_root] lint. This lint prevents data of the marked type from being used on the stack. See the source for more details.

Macros

symbols

Structs

FnDefVisitor

Symbols

UnrootedPass

Lint for ensuring safe usage of unrooted pointers

Statics

UNROOTED_MUST_ROOT

Functions

__rustc_plugin_registrar

has_lint_attr

in_derive_expn

is_unrooted_ty

Checks if a type is unrooted or contains any owned unrooted types

match_def_path

check if a DefId’s path matches the given absolute type path usage e.g. with match_def_path(cx, id, &["core", "option", "Option"])

registrar