Struct script_plugins::UnrootedPass

pub(crate) struct UnrootedPass {
    pub(crate) symbols: Symbols,
}

Lint for ensuring safe usage of unrooted pointers

This lint (disable with -A unrooted-must-root/#[allow(unrooted_must_root)]) ensures that #[unrooted_must_root_lint::must_root] values are used correctly.

“Incorrect” usage includes:

Not being used in a struct/enum field which is not #[unrooted_must_root_lint::must_root] itself

Not being used as an argument to a function (Except onces named new and new_inherited)

Not being bound locally in a let statement, assignment, for loop, or match statement.

This helps catch most situations where pointers like JS<T> are used in a way that they can be invalidated by a GC pass.

Structs which have their own mechanism of rooting their unrooted contents (e.g. ScriptThread) can be marked as #[allow(unrooted_must_root)]. Smart pointers which root their interior type can be marked as #[unrooted_must_root_lint::allow_unrooted_interior]

Fields
symbols: Symbols
Implementations

impl UnrootedPass

pub fn new(symbols: Symbols) -> UnrootedPass